[#199] refactor : 타입,품종 컴포넌트 분리
