+++ app/_components/PetRegister/index.tsx
@@ -2,15 +2,13 @@
 
 import { SubmitHandler, useForm } from "react-hook-form";
 import { PET_NAME_RULES, PET_WEIGHT_RULES, PET_REGISTERNUMBER_RULES, PET_PLACEHOLDER, PET_GENDER_RULES } from "@/app/_constants/inputConstant";
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import * as styles from "./style.css";
 import DefaultImage from "@/public/images/pet-profile-default.svg?url";
 import cameraIcon from "@/public/icons/camera.svg?url";
 import Image from "next/image";
 import PetDateInput from "@/app/_components/PetRegister/component/PetdateInput";
-import { petOptions } from "@/public/data/petOptions";
 import ErrorMessage from "@/app/_components/ErrorMessage";
-import DropdownIcon from "@/public/icons/drop-down-icon.svg";
 import OptionalMessage from "./component/OptionalCheck";
 import CloseIcon from "@/public/icons/close.svg?url";
 import BackIcon from "@/public/icons/chevron-left.svg?url";
@@ -18,6 +16,7 @@ import { useRouter } from "next/navigation";
 import { postPet } from "@/app/_api/pets";
 import { useModal } from "@/app/_hooks/useModal";
 import Modal from "@/app/_components/Modal";
+import TypeBreedSelection from "@/app/_components/PetRegister/component/TypeBreedSelection";
 
 export interface IFormInput {
   petName: string;
@@ -38,11 +37,6 @@ const PetRegister = () => {
   const { isModalOpen, openModalFunc, closeModalFunc } = useModal();
   const [profileImage, setProfileImage] = useState<File | string | null>(DefaultImage);
   const [section, setSection] = useState(1);
-  const [breedOpen, setBreedOpen] = useState(false); //모달상태
-  const [typeOpen, setTypeOpen] = useState(false); //모달상태
-  const dropdownRef = useRef<HTMLUListElement>(null); //모달 외부 클릭시 닫히도록
-  const [selectedType, setSelectedType] = useState(""); //타입 선택 반영
-  const [selectedBreed, setSelectedBreed] = useState(""); //품종 선택 반영
   const [selectedGender, setSelectedGender] = useState<string>(""); //성별 선택 반영
   const [selectedNeutering, setSelectedNeutering] = useState(""); //중성화 선택 반영
   const [isWeightDisabled, setIsWeightDisabled] = useState(false); //몸무게 모르겠어요 반영
@@ -121,39 +115,6 @@ const PetRegister = () => {
     setProfileImage(watch("image") || DefaultImage);
   }, [watch]);
 
-  //드롭다운 외부 클릭시 닫히게 하기
-  useEffect(() => {
-    function handleClickOutside(event: MouseEvent) {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target as Node)) {
-        setBreedOpen(false);
-        setTypeOpen(false);
-      }
-    }
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [dropdownRef]);
-
-  //드롭다운
-  useEffect(() => {
-    setSelectedBreed("");
-    setValue("breed", "");
-  }, [selectedType, setValue]);
-
-  const handleTypeClick = (type: string) => {
-    setValue("type", type);
-    setSelectedType(type);
-    setSelectedBreed("");
-    setTypeOpen((prev) => !prev);
-  };
-
-  const handleBreedClick = (breed: string) => {
-    setValue("breed", breed);
-    setSelectedBreed(breed);
-    setBreedOpen((prev) => !prev);
-  };
-
   //중성화
   const handleNeuteringChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setSelectedNeutering(e.target.value);
@@ -198,58 +159,7 @@ const PetRegister = () => {
       <input className={styles.writeInput} {...register("petName", PET_NAME_RULES)} placeholder={PET_PLACEHOLDER.name} />
       {errors.petName && <ErrorMessage message={errors.petName.message} />}
 
-      {/* 타입 */}
-      <label className={styles.label}>타입*</label>
-      <div>
-        <button className={`${styles.selectBox} ${typeOpen ? styles.selectBoxOpen : ""}`} onClick={() => setTypeOpen(true)}>
-          {selectedType || "타입을 선택하세요"}
-          <DropdownIcon className={`${styles.dropdownIcon} ${typeOpen ? styles.dropdownIconOpen : ""}`} />
-        </button>
-        {typeOpen && (
-          <ul className={styles.optionsList} ref={dropdownRef}>
-            {Object.keys(petOptions).map((type: string, index: number) => (
-              <li key={index} value={type}>
-                <button type="button" className={styles.optionButton} onClick={() => handleTypeClick(type)} {...register("type")}>
-                  {type}
-                </button>
-              </li>
-            ))}
-          </ul>
-        )}
-      </div>
-
-      {/* 품종 */}
-      <label className={styles.label}>품종*</label>
-      {selectedType !== "기타" && (
-        <button className={`${styles.selectBox} ${breedOpen ? styles.selectBoxOpen : ""}`} onClick={() => setBreedOpen(!breedOpen)}>
-          {selectedBreed || "품종을 선택하세요"}
-          <DropdownIcon className={`${styles.dropdownIcon} ${breedOpen ? styles.dropdownIconOpen : ""}`} />
-        </button>
-      )}
-      {breedOpen && selectedType !== "기타" && (
-        <ul className={styles.optionsList} ref={dropdownRef}>
-          {petOptions[selectedType]?.map((breed: string, index: number) => (
-            <li key={index} value={breed}>
-              <button type="button" className={styles.optionButton} onClick={() => handleBreedClick(breed)} {...register("breed")}>
-                {breed}
-              </button>
-            </li>
-          ))}
-        </ul>
-      )}
-
-      {selectedType === "기타" && (
-        <>
-          <input
-            className={styles.writeInput}
-            placeholder="품종을 직접 입력하세요"
-            {...register("breed", {
-              required: "내용을 입력해주세요",
-            })}
-            autoFocus
-          />
-        </>
-      )}
+      <TypeBreedSelection register={register} setValue={setValue} />
 
       <button className={styles.button} onClick={handleNextSection} disabled={!isSectionValid}>
         다음
